pyproject.toml
[tool.poetry]
name = "homework-11"
version = "0.1.0"
description = ""
authors = ["VlodyaKr <vlodya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
mariadb = "^1.1.4"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
python-dotenv = "^0.21.0"
marshmallow = "^3.18.0"
bcrypt = "^4.0.0"
aioshutil = "^1.1"
pydantic = {extras = ["email"], version = "^1.10.2"}
flask-marshmallow = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
